% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ZIM.R
\docType{data}
\name{data_ZIM}
\alias{data_ZIM}
\title{Simulated microbiome data}
\format{
A data frame with 300 observations and 10 taxons:
\describe{
\item{x}{The independent variable}
\item{y1}{The outcome variable}
\item{libSize}{The library size}
\item{taxon}{The relative abundance of taxons}
}
}
\usage{
data_ZIM
}
\description{
The Simulated microbiome dataset to show the use of \code{ZIM_func}.
}
\keyword{datasets}
