% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedZIM_func.R
\name{MedZIM_func}
\alias{MedZIM_func}
\title{Perform MedZIM model}
\usage{
MedZIM_func(
  dat,
  xVar,
  yVar,
  taxon_name,
  libSize_name,
  obs_gt_0 = 2,
  obs_eq_0 = 2,
  inter_x_mg0 = T,
  inter_x_m = F,
  eval.max = 200,
  iter.max = 200,
  x_from = 0,
  x_to = 1,
  type1error = 0.05,
  paraJobs = 4
)
}
\arguments{
\item{dat}{The dataset used for the model.}

\item{xVar}{The name of X (independent variable) in the data.}

\item{yVar}{The name of Y (outcome variable) in the data.}

\item{taxon_name}{The string that can identify the taxon variables.}

\item{libSize_name}{The name of library size variable in the data.}

\item{obs_gt_0}{The at least number of non-zero observations that a taxon should have to be included into the analysis.}

\item{obs_eq_0}{The at least number of zero observations that a taxon should have to be included in group 1.}

\item{inter_x_mg0}{Whether to include the \eqn{\beta4} term into the model.}

\item{inter_x_m}{Whether to include the \eqn{\beta5} term into the model.}

\item{eval.max}{Maximum number of evaluations of objective functions allowed in nlminb function.}

\item{iter.max}{Maximum number of iterations allowed in nlminb function.}

\item{x_from}{The value X starts from in evaluating the effect of X.}

\item{x_to}{The value X ends to in evaluating the effect of X.}

\item{type1error}{The type I error.}

\item{paraJobs}{The number of cores for parallel computing.}
}
\value{
A list contains all results.
\item{fullList}{A list contains resutls for each taxon. See details.}
\item{validTaxaNames}{The taxons that were in group 1, which had enough zero and non-zero observations.}
\item{continuTaxa}{The taxons that were in group 2, which only had non-zero observations.}
}
\description{
{
The Outcome model is:
\deqn{Y=\beta0+\beta1M+\beta2 1_{(M>0)}+\beta3X+\beta4X1_{(M>0)}+\beta5XM+\epsilon}
}
}
\details{
The output list contains information of all taxons to be analyzed for mediation effects. Each taxon result is also
a list, with first object contains p value, estimates, and CI of NIE1, NIE2, NIE, and each parameters.
}
\examples{
MedZIM_func(dat=data_ZIM, xVar = "x",yVar = "y1",
taxon_name = "taxon",libSize_name = "libSize",paraJobs=2)


}
